feat: tabs on code mgmt page with code requests
