+++ src/components/EnterpriseApp/constants.js
@@ -4,4 +4,5 @@ export const ROUTE_NAMES = {
   bulkEnrollment: 'enrollment',
   subscriptionManagement: 'subscriptions',
   settings: 'settings',
+  codeManagement: 'coupons',
 };
