--- src/containers/CodeManagementPage/data/constants.js
@@ -1,7 +1,0 @@
-export const MANAGE_CODES_TAB = 'coupons';
-export const MANAGE_REQUESTS_TAB = 'requests';
-
-export const MANAGE_CODES_LABEL = 'Manage Codes';
-export const MANAGE_REQUESTS_LABEL = 'Manage Requests';
-
-export const DEFAULT_TAB = MANAGE_CODES_TAB;
