+++ src/data/services/EnterpriseAccessApiService.js
@@ -81,11 +81,19 @@ class EnterpriseAccessApiService {
     return EnterpriseAccessApiService.apiClient().get(url);
   }
 
-  static getCouponCodeRequests(enterpriseId, options = {}) {
+  static getCouponCodeRequests(enterpriseId, requestStatusFilters, options = {}) {
     const params = new URLSearchParams({
       enterprise_customer_uuid: enterpriseId,
       ...options,
     });
+    if (requestStatusFilters) {
+      // TODO: the API doesn't seem to handle multiple state filters at once, returning results only
+      // for the latest state query parameter passed to the URL. Also tried using comma-separated list
+      // which returns a 400.
+      requestStatusFilters.forEach((requestStatus) => {
+        params.append('state', requestStatus);
+      });
+    }
     const url = `${EnterpriseAccessApiService.baseUrl}/coupon-code-requests/?${params.toString()}`;
     return EnterpriseAccessApiService.apiClient().get(url);
   }
